Create README.md
I want to bring a Team of people wanting to help and build a simple app to help an solve a massive problem.
I am from México so i will be speaking on spanish from time to time.

They are several problems that mexican farmers have to face every day. Low income, low resources, and they do not have 
the possibility to sell they products at a good price. They lack transportation, media exposure and the most important of all
chance to acess a good education. There is a lot of people that abuse this and buy all the production of one farmer for a irreal price,
then re sell them or buy all the product available to get the prices to raise. It an injustice and the government wont do anything
really aiming to solve it, because they got interest conflicts. So i want to solve it, that is why i need your help. If i go to the
authorities of education, investigation or pretty any organitazion on my country with this proposal they would laught at me. In fact
they already launched an app with this format, but they didnt make it easy to access, the UI its horrible, they didnt attend the users
complains. So they said they made an app to resolve the problem but it didnt work.
I want them to take them by surprise, so i want the app to get build outside possible obstructors.
Ill use all the resource that student developer pack make avaible. 
I do not have any programing skills. But i know the field, the people, and the market. 
There are a lot of more troubles but i want to focus in opportunities to acces to the market.
I want to make an app that every farmer could have on a smarthpone.
i know that projects have to be sliced to the minimun viable product so lets focus on Avocado for start.
An app made for the little produces to make public all the product they have to sell so anyone interested could buy the product through
the app and give choices for transportation of the product, the buyer absorb the cost of transportation by contracting a third party
service. or another options. I know there are a lot of market app outside there. But the most popular ones do not have a specific audience,
so if a farmer post a product hopping to find a buyer, the publication gets buried under all the other most profesional publications.

So, the goal. Make it free to use for everyone. tips for making they product more attractive. Let them include photos and clips. 
Let them do short streams so they can show they product and where it came frome.

Idiomas: Why those languages? 
Recently Mexico and UE reach an agreement that makes easier to trade. That agreement open a huge market
for these kind of product to be sell. I agree that open trade chances are oportunities for a country to grow, but that case is true
if there are no corruption in the country.
First step: Get going.
Would you help me?

Aguacate - Avocado